requirements.txt
langchain
langchain-community
langchain_groq
chromadb
streamlit
python-dotenv
sentence-transformers
langchain_huggingface